fish_pipeline_v5/preprocessing.py
import os
import cv2
import numpy as np
from pathlib import Path
from PIL import Image, ImageEnhance, ImageFilter
import albumentations as A
import json

# Load config
with open('config.json') as f:
    config = json.load(f)

DATASET_DIR = config['paths']['dataset']
OUTPUT_DIR = os.path.join(config['paths']['output'], 'preprocessed')
TARGET_SIZE = config['preprocessing']['target_size']

PREPROCESSING_STRATEGIES = [
    {"name": "original_enhanced", "params": {"enhance": True}},
    {"name": "bright_contrast", "params": {"brightness": 1.3, "contrast": 1.2}},
    {"name": "sharp_saturation", "params": {"sharpness": 1.5, "saturation": 1.3}},
    {"name": "clahe_gamma", "params": {"clahe": 2.0, "gamma": 1.2}},
    {"name": "histogram_eq", "params": {"hist_eq": True, "brightness": 1.1}},
    {"name": "blur_denoise", "params": {"gaussian_blur": 0.5, "median_filter": True}},
    {"name": "edge_enhance", "params": {"edge_enhance": True, "sharpness": 2.0}},
    {"name": "color_balance", "params": {"auto_contrast": True, "color": 1.4}},
    {"name": "lighting_fix", "params": {"brightness": 0.8, "gamma": 0.9}},
    {"name": "detail_enhance", "params": {"unsharp_mask": True, "contrast": 1.4}}
]

def preprocess_image(image_path, output_dir):
    image = cv2.imread(image_path)
    if image is None:
        return []
    base_name = Path(image_path).stem
    results = []
    for i, strategy in enumerate(PREPROCESSING_STRATEGIES):
        processed = image.copy()
        # ...apply each strategy (implement as needed)...
        out_path = os.path.join(output_dir, f"{base_name}_{i+1:02d}_{strategy['name']}.jpg")
        cv2.imwrite(out_path, processed)
        results.append(out_path)
    return results

def preprocess_dataset():
    os.makedirs(OUTPUT_DIR, exist_ok=True)
    for species in os.listdir(DATASET_DIR):
        species_dir = os.path.join(DATASET_DIR, species)
        if not os.path.isdir(species_dir):
            continue
        out_species_dir = os.path.join(OUTPUT_DIR, species)
        os.makedirs(out_species_dir, exist_ok=True)
        for img_file in os.listdir(species_dir):
            img_path = os.path.join(species_dir, img_file)
            preprocess_image(img_path, out_species_dir)

if __name__ == "__main__":
    preprocess_dataset()
