fish_pipeline_v5/requirements.txt
torch
opencv-python
albumentations
ultralytics
scikit-learn
numpy
Pillow
matplotlib
